requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.9.3
pandas>=1.3.0
numpy>=1.21.0,<2.0.0
scikit-learn>=0.24.2
nltk>=3.6.2
spacy>=3.1.0,<3.2.0
google-api-python-client>=2.0.0
python-multipart>=0.0.5
openai>=0.27.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Development
black>=21.7b0
flake8>=3.9.2
isort>=5.9.3 